+++ src/main/java/org/apache/parquet/hadoop/metadata/IndexedParquetMetadata.java
@@ -1,20 +1,20 @@
 package org.apache.parquet.hadoop.metadata;
 
 
-import org.apache.parquet.it.unimi.dsi.fastutil.longs.LongList;
+import org.apache.parquet.it.unimi.dsi.fastutil.ints.IntList;
 
 import java.util.List;
 
 public class IndexedParquetMetadata extends ParquetMetadata {
 
-    private List<LongList> rowIdsList;
+    private List<IntList> rowIdsList;
 
-    public IndexedParquetMetadata(FileMetaData fileMetaData, List<BlockMetaData> blocks, List<LongList> rowIdsList) {
+    public IndexedParquetMetadata(FileMetaData fileMetaData, List<BlockMetaData> blocks, List<IntList> rowIdsList) {
         super(fileMetaData, blocks);
         this.rowIdsList = rowIdsList;
     }
 
-    public List<LongList> getRowIdsList() {
+    public List<IntList> getRowIdsList() {
         return rowIdsList;
     }
 }
